Adding the server ping to check

+++ Client.py
@@ -1,8 +1,9 @@
 import socket
-import sys
 import os
 from threading import Thread
 import time
+import subprocess
+import requests
 from pyautogui import press, typewrite, hotkey
 import requests
 
@@ -13,6 +14,8 @@ global break_all
 global new_IP
 global Ip
 
+FNULL = open(os.devnull, 'w')
+
 break_all = False
 new_IP = False
 
@@ -22,10 +25,13 @@ class starter(Thread):
     global new_IP
     global break_all
     global Ip
+    global connected
 
 
 
     def __init__(self):
+        global soc
+        global connected
         connected = False
         global break_all
         global Ip
@@ -47,7 +53,6 @@ class starter(Thread):
                 host = Q
                 try:
                     soc.connect((host, port))
-                    self.soc = soc
                     print("Success connecting to server")
                     message = socket.gethostname()
                     print(message)
@@ -68,7 +73,6 @@ class starter(Thread):
             while connected == False:
                 try:
                     soc.connect((host, port))
-                    self.soc = soc
                     print("Success connecting to server")
                     message = socket.gethostname()
                     print(message)
@@ -81,6 +85,7 @@ class starter(Thread):
                     pass
 
     def run(self):
+        global soc
         while self.running:
             if break_all == True:
                 return
@@ -101,18 +106,19 @@ class starter(Thread):
                 elif message == "/rm":
                     rm_message = '--RM--'
                     soc.sendall(rm_message.encode("utf8"))
-                    rm(self)
+                    rm()
             print('sending quit')
             message = "--quit--"
             soc.send(message.encode("utf-8"))
             time.sleep(1)
             return message
 
-    def rm(self):
-        message = ""
-        while message != "/back":
-            message = input(" -> ")
-            soc.sendall(message.encode("utf8"))
+
+def rm():
+    message = ""
+    while message != "/back":
+        message = input(" -> ")
+        soc.sendall(message.encode("utf8"))
 
 class Receive(starter):
     global break_all
@@ -124,7 +130,6 @@ class Receive(starter):
         global runnin
         Thread.__init__(self)
         runnin = starter.runnin
-        soc = starter.soc
 
     def run(self):
         global break_all
@@ -132,17 +137,18 @@ class Receive(starter):
         while True:
             try:
                 data = soc.recv(1024).decode()
-                if not str(data) == ("test"):
+                if not str(data).__contains__("--TEST--"):
                     if not str(data).__contains__('testtest'):
                         if data == "":
                             break_all = True
                             return
-                        if str(data) == "--SENDING_FILE--":
+                        elif str(data) == "--SENDING_FILE--":
                             print('recieving file...')
                             os.system('Get.py')
                         else:
                             print('\n' + "Recieving message from server: " + data)
                             time.sleep(.1)
+
             except:
                 print('server closed')
                 break_all = True
@@ -150,18 +156,37 @@ class Receive(starter):
                 os._exit(1)
                 return
 
+
 class Checker(Thread):
+    global soc
+    global Ip
+
     def __init__(self):
+        global soc
+        global Ip
         Thread.__init__(self)
 
     def run(self):
-        print('starting internet check')
+        global soc
+        global Ip
+        print('starting')
+        with open("ping.bat", "w+") as ping:
+            ping.write("ping.exe " + Ip + " -n 2 > ping.txt")
         while True:
-            try:
-                requests.get('http://216.58.192.142', timeout=1)
-            except:
-                os.remove('tmp.txt')
-                os._exit(1)
+            print('starting internet check')
+            press('enter')
+            while True:
+                subprocess.run("ping.bat", stdout=FNULL)
+                with open("ping.txt", "r") as file:
+                    file = file.read()
+                if file.__contains__("Destination host unreachable."):
+                    print('server closed')
+                    break_all = True
+                    os.remove("tmp.txt")
+                    os._exit(1)
+                    return
+                else:
+                    pass
 
 try:
     with open("IP.txt","r") as IP:
